requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
huggingface_hub>=0.19.0
